testing for TypeFactory
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@1865 23e1bc29-d954-0410-9766-47de34158ba0

+++ ConstraintEngine/component/IntervalTypeFactory.cc
@@ -39,7 +39,7 @@ namespace Prototype {
 
   double IntervalTypeFactory::createValue(std::string value) const
   {
-    return atoi(value.c_str());
+    return atof(value.c_str());
   }
 
 } // namespace Prototype
